Build with Hugo 0.50
This means the previous `---` in the 'Making a server - the plan' article has
become just `-` which looks nicer. I have added an alias to the article for the
old URL to compensate for this - however, the aliases doesn't actually make it
through to the optimised build as the pages aren't referenced from anywhere.
That's something to look into for another day.

I have added a couple of extra flags to the Assetgraph Builder command, which
will make it quit in case a warning happens, but also make it add subresource
integrity tags to any resources it can add it for, and lastly to make it
optimise images automatically.